requirements.txt
pandas==2.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
duckdb==1.1.3
pyarrow==17.0.0
